clean up inner classes
